@@ -4,7 +4,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -12,9 +12,10 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  font_awesome_flutter: ^8.8.1
-  smooth_star_rating: ^1.1.1
+  cupertino_icons: ^1.0.2
+  font_awesome_flutter: ^10.1.0
+  flutter_rating_bar: ^4.0.1
+
   intl: ^0.17.0
   fair:
     path: ../../fair
