@@ -1,8 +1,8 @@
 import 'package:best_flutter_ui_templates/fitness_app/models/tabIcon_data.dart';
-import 'package:best_flutter_ui_templates/fitness_app/traning/training_screen.dart';
+import 'package:best_flutter_ui_templates/fitness_app/training/training_screen.dart';
 import 'package:flutter/material.dart';
 import 'bottom_navigation_view/bottom_bar_view.dart';
-import 'fintness_app_theme.dart';
+import 'fitness_app_theme.dart';
 import 'my_diary/my_diary_screen.dart';
 
 class FitnessAppHomeScreen extends StatefulWidget {
@@ -12,7 +12,7 @@ class FitnessAppHomeScreen extends StatefulWidget {
 
 class _FitnessAppHomeScreenState extends State<FitnessAppHomeScreen>
     with TickerProviderStateMixin {
-  AnimationController animationController;
+  AnimationController? animationController;
 
   List<TabIconData> tabIconsList = TabIconData.tabIconsList;
 
@@ -35,7 +35,7 @@ class _FitnessAppHomeScreenState extends State<FitnessAppHomeScreen>
 
   @override
   void dispose() {
-    animationController.dispose();
+    animationController?.dispose();
     super.dispose();
   }
 
@@ -80,7 +80,7 @@ class _FitnessAppHomeScreenState extends State<FitnessAppHomeScreen>
           addClick: () {},
           changeIndex: (int index) {
             if (index == 0 || index == 2) {
-              animationController.reverse().then<dynamic>((data) {
+              animationController?.reverse().then<dynamic>((data) {
                 if (!mounted) {
                   return;
                 }
@@ -90,7 +90,7 @@ class _FitnessAppHomeScreenState extends State<FitnessAppHomeScreen>
                 });
               });
             } else if (index == 1 || index == 3) {
-              animationController.reverse().then<dynamic>((data) {
+              animationController?.reverse().then<dynamic>((data) {
                 if (!mounted) {
                   return;
                 }
