@@ -1,15 +1,15 @@
 import 'package:best_flutter_ui_templates/fitness_app/ui_view/wave_view.dart';
-import 'package:best_flutter_ui_templates/fitness_app/fintness_app_theme.dart';
+import 'package:best_flutter_ui_templates/fitness_app/fitness_app_theme.dart';
 import 'package:best_flutter_ui_templates/main.dart';
 import 'package:flutter/material.dart';
 
 class WaterView extends StatefulWidget {
   const WaterView(
-      {Key key, this.mainScreenAnimationController, this.mainScreenAnimation})
+      {Key? key, this.mainScreenAnimationController, this.mainScreenAnimation})
       : super(key: key);
 
-  final AnimationController mainScreenAnimationController;
-  final Animation<dynamic> mainScreenAnimation;
+  final AnimationController? mainScreenAnimationController;
+  final Animation<double>? mainScreenAnimation;
 
   @override
   _WaterViewState createState() => _WaterViewState();
@@ -24,13 +24,13 @@ class _WaterViewState extends State<WaterView> with TickerProviderStateMixin {
   @override
   Widget build(BuildContext context) {
     return AnimatedBuilder(
-      animation: widget.mainScreenAnimationController,
-      builder: (BuildContext context, Widget child) {
+      animation: widget.mainScreenAnimationController!,
+      builder: (BuildContext context, Widget? child) {
         return FadeTransition(
-          opacity: widget.mainScreenAnimation,
+          opacity: widget.mainScreenAnimation!,
           child: Transform(
             transform: Matrix4.translationValues(
-                0.0, 30 * (1.0 - widget.mainScreenAnimation.value), 0.0),
+                0.0, 30 * (1.0 - widget.mainScreenAnimation!.value), 0.0),
             child: Padding(
               padding: const EdgeInsets.only(
                   left: 24, right: 24, top: 16, bottom: 18),
