@@ -1,24 +1,24 @@
 import 'package:best_flutter_ui_templates/main.dart';
 import 'package:flutter/material.dart';
-import '../fintness_app_theme.dart';
+import '../fitness_app_theme.dart';
 
 class WorkoutView extends StatelessWidget {
-  final AnimationController animationController;
-  final Animation animation;
+  final AnimationController? animationController;
+  final Animation<double>? animation;
 
-  const WorkoutView({Key key, this.animationController, this.animation})
+  const WorkoutView({Key? key, this.animationController, this.animation})
       : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return AnimatedBuilder(
-      animation: animationController,
-      builder: (BuildContext context, Widget child) {
+      animation: animationController!,
+      builder: (BuildContext context, Widget? child) {
         return FadeTransition(
-          opacity: animation,
+          opacity: animation!,
           child: new Transform(
             transform: new Matrix4.translationValues(
-                0.0, 30 * (1.0 - animation.value), 0.0),
+                0.0, 30 * (1.0 - animation!.value), 0.0),
             child: Padding(
               padding: const EdgeInsets.only(
                   left: 24, right: 24, top: 16, bottom: 18),
@@ -59,7 +59,7 @@ class WorkoutView extends StatelessWidget {
                       ),
                       Padding(
                         padding: const EdgeInsets.only(top: 8.0),
-                        child: Text(
+                        child: const Text(
                           'Legs Toning and\nGlutes Workout at Home',
                           textAlign: TextAlign.left,
                           style: TextStyle(
@@ -90,7 +90,7 @@ class WorkoutView extends StatelessWidget {
                             ),
                             Padding(
                               padding: const EdgeInsets.only(left: 4.0),
-                              child: Text(
+                              child: const Text(
                                 '68 min',
                                 textAlign: TextAlign.center,
                                 style: TextStyle(
