@@ -1,14 +1,13 @@
-import 'package:fair/fair.dart';
 import 'package:flutter/material.dart';
 
 import 'hotel_app_theme.dart';
 
 class SliderView extends StatefulWidget {
-  const SliderView({Key key, this.onChangedistValue, this.distValue})
+  const SliderView({Key? key, this.onChangedistValue, this.distValue})
       : super(key: key);
 
-  final Function(double) onChangedistValue;
-  final double distValue;
+  final Function(double)? onChangedistValue;
+  final double? distValue;
 
   @override
   _SliderViewState createState() => _SliderViewState();
@@ -19,7 +18,7 @@ class _SliderViewState extends State<SliderView> {
 
   @override
   void initState() {
-    distValue = widget.distValue;
+    distValue = widget.distValue!;
     super.initState();
   }
 
@@ -57,7 +56,7 @@ class _SliderViewState extends State<SliderView> {
                   distValue = value;
                 });
                 try {
-                  widget.onChangedistValue(distValue);
+                  widget.onChangedistValue!(distValue);
                 } catch (_) {}
               },
               min: 0,
@@ -92,23 +91,23 @@ class CustomThumbShape extends SliderComponentShape {
 
   @override
   void paint(
-    PaintingContext context,
-    Offset thumbCenter, {
-    Animation<double> activationAnimation,
-    Animation<double> enableAnimation,
-    bool isDiscrete,
-    TextPainter labelPainter,
-    RenderBox parentBox,
-    Size sizeWithOverflow,
-    SliderThemeData sliderTheme,
-    TextDirection textDirection,
-    double textScaleFactor,
-    double value,
-  }) {
+      PaintingContext context,
+      Offset thumbCenter, {
+        Animation<double>? activationAnimation,
+        Animation<double>? enableAnimation,
+        bool? isDiscrete,
+        TextPainter? labelPainter,
+        RenderBox? parentBox,
+        Size? sizeWithOverflow,
+        SliderThemeData? sliderTheme,
+        TextDirection textDirection = TextDirection.ltr,
+        double? textScaleFactor,
+        double? value,
+      }) {
     final Canvas canvas = context.canvas;
     final ColorTween colorTween = ColorTween(
-      begin: sliderTheme.disabledThumbColor,
-      end: sliderTheme.thumbColor,
+      begin: sliderTheme?.disabledThumbColor,
+      end: sliderTheme?.thumbColor,
     );
     canvas.drawPath(
         Path()
@@ -119,13 +118,13 @@ class CustomThumbShape extends SliderComponentShape {
         Paint()
           ..color = Colors.black.withOpacity(0.5)
           ..maskFilter =
-              MaskFilter.blur(BlurStyle.normal, convertRadiusToSigma(8)));
+          MaskFilter.blur(BlurStyle.normal, convertRadiusToSigma(8)));
 
     final Paint cPaint = Paint();
     cPaint..color = Colors.white;
     cPaint..strokeWidth = 14 / 2;
     canvas.drawCircle(Offset(thumbCenter.dx, thumbCenter.dy), 12, cPaint);
-    cPaint..color = colorTween.evaluate(enableAnimation);
+    cPaint..color = colorTween.evaluate(enableAnimation!)!;
     canvas.drawCircle(Offset(thumbCenter.dx, thumbCenter.dy), 10, cPaint);
   }
 
