@@ -10,20 +10,20 @@ class HomeListProxy extends FairDelegate {
     this.callBack,
   });
 
-  final HomeList listData;
-  final VoidCallback callBack;
+  final HomeList? listData;
+  final VoidCallback? callBack;
 
   @override
   Map<String, PropertyValue> bindValue() {
     var value = super.bindValue();
-    value['imagePath'] = () => listData.imagePath;
+    value['imagePath'] = () => listData!.imagePath;
     return value;
   }
 
   @override
   Map<String, Function> bindFunction() {
     var fun = super.bindFunction();
-    fun['callBack'] = callBack;
+    fun['callBack'] = callBack!;
     return fun;
   }
 }
