@@ -43,7 +43,7 @@ class MediterranesnDietView extends StatelessWidget {
                   children: <Widget>[
                     Padding(
                       padding:
-                      const EdgeInsets.only(top: 16, left: 16, right: 16),
+                          const EdgeInsets.only(top: 16, left: 16, right: 16),
                       child: Row(
                         children: <Widget>[
                           Expanded(
@@ -68,9 +68,9 @@ class MediterranesnDietView extends StatelessWidget {
                                         padding: const EdgeInsets.all(8.0),
                                         child: Column(
                                           mainAxisAlignment:
-                                          MainAxisAlignment.center,
+                                              MainAxisAlignment.center,
                                           crossAxisAlignment:
-                                          CrossAxisAlignment.start,
+                                              CrossAxisAlignment.start,
                                           children: <Widget>[
                                             Padding(
                                               padding: const EdgeInsets.only(
@@ -80,7 +80,7 @@ class MediterranesnDietView extends StatelessWidget {
                                                 textAlign: TextAlign.center,
                                                 style: TextStyle(
                                                   fontFamily:
-                                                  FitnessAppTheme.fontName,
+                                                      FitnessAppTheme.fontName,
                                                   fontWeight: FontWeight.w500,
                                                   fontSize: 16,
                                                   letterSpacing: -0.1,
@@ -91,9 +91,9 @@ class MediterranesnDietView extends StatelessWidget {
                                             ),
                                             Row(
                                               mainAxisAlignment:
-                                              MainAxisAlignment.center,
+                                                  MainAxisAlignment.center,
                                               crossAxisAlignment:
-                                              CrossAxisAlignment.end,
+                                                  CrossAxisAlignment.end,
                                               children: <Widget>[
                                                 SizedBox(
                                                   width: 28,
@@ -103,17 +103,17 @@ class MediterranesnDietView extends StatelessWidget {
                                                 ),
                                                 Padding(
                                                   padding:
-                                                  const EdgeInsets.only(
-                                                      left: 4, bottom: 3),
+                                                      const EdgeInsets.only(
+                                                          left: 4, bottom: 3),
                                                   child: Text(
                                                     '${(1127 * animation!.value).toInt()}',
                                                     textAlign: TextAlign.center,
                                                     style: TextStyle(
                                                       fontFamily:
-                                                      FitnessAppTheme
-                                                          .fontName,
+                                                          FitnessAppTheme
+                                                              .fontName,
                                                       fontWeight:
-                                                      FontWeight.w600,
+                                                          FontWeight.w600,
                                                       fontSize: 16,
                                                       color: FitnessAppTheme
                                                           .darkerText,
@@ -122,17 +122,17 @@ class MediterranesnDietView extends StatelessWidget {
                                                 ),
                                                 Padding(
                                                   padding:
-                                                  const EdgeInsets.only(
-                                                      left: 4, bottom: 3),
+                                                      const EdgeInsets.only(
+                                                          left: 4, bottom: 3),
                                                   child: Text(
                                                     'Kcal',
                                                     textAlign: TextAlign.center,
                                                     style: TextStyle(
                                                       fontFamily:
-                                                      FitnessAppTheme
-                                                          .fontName,
+                                                          FitnessAppTheme
+                                                              .fontName,
                                                       fontWeight:
-                                                      FontWeight.w600,
+                                                          FontWeight.w600,
                                                       fontSize: 12,
                                                       letterSpacing: -0.2,
                                                       color: FitnessAppTheme
@@ -167,9 +167,9 @@ class MediterranesnDietView extends StatelessWidget {
                                         padding: const EdgeInsets.all(8.0),
                                         child: Column(
                                           mainAxisAlignment:
-                                          MainAxisAlignment.center,
+                                              MainAxisAlignment.center,
                                           crossAxisAlignment:
-                                          CrossAxisAlignment.start,
+                                              CrossAxisAlignment.start,
                                           children: <Widget>[
                                             Padding(
                                               padding: const EdgeInsets.only(
@@ -179,7 +179,7 @@ class MediterranesnDietView extends StatelessWidget {
                                                 textAlign: TextAlign.center,
                                                 style: TextStyle(
                                                   fontFamily:
-                                                  FitnessAppTheme.fontName,
+                                                      FitnessAppTheme.fontName,
                                                   fontWeight: FontWeight.w500,
                                                   fontSize: 16,
                                                   letterSpacing: -0.1,
@@ -190,9 +190,9 @@ class MediterranesnDietView extends StatelessWidget {
                                             ),
                                             Row(
                                               mainAxisAlignment:
-                                              MainAxisAlignment.center,
+                                                  MainAxisAlignment.center,
                                               crossAxisAlignment:
-                                              CrossAxisAlignment.end,
+                                                  CrossAxisAlignment.end,
                                               children: <Widget>[
                                                 SizedBox(
                                                   width: 28,
@@ -202,17 +202,17 @@ class MediterranesnDietView extends StatelessWidget {
                                                 ),
                                                 Padding(
                                                   padding:
-                                                  const EdgeInsets.only(
-                                                      left: 4, bottom: 3),
+                                                      const EdgeInsets.only(
+                                                          left: 4, bottom: 3),
                                                   child: Text(
                                                     '${(102 * animation!.value).toInt()}',
                                                     textAlign: TextAlign.center,
                                                     style: TextStyle(
                                                       fontFamily:
-                                                      FitnessAppTheme
-                                                          .fontName,
+                                                          FitnessAppTheme
+                                                              .fontName,
                                                       fontWeight:
-                                                      FontWeight.w600,
+                                                          FontWeight.w600,
                                                       fontSize: 16,
                                                       color: FitnessAppTheme
                                                           .darkerText,
@@ -221,17 +221,17 @@ class MediterranesnDietView extends StatelessWidget {
                                                 ),
                                                 Padding(
                                                   padding:
-                                                  const EdgeInsets.only(
-                                                      left: 8, bottom: 3),
+                                                      const EdgeInsets.only(
+                                                          left: 8, bottom: 3),
                                                   child: Text(
                                                     'Kcal',
                                                     textAlign: TextAlign.center,
                                                     style: TextStyle(
                                                       fontFamily:
-                                                      FitnessAppTheme
-                                                          .fontName,
+                                                          FitnessAppTheme
+                                                              .fontName,
                                                       fontWeight:
-                                                      FontWeight.w600,
+                                                          FontWeight.w600,
                                                       fontSize: 12,
                                                       letterSpacing: -0.2,
                                                       color: FitnessAppTheme
@@ -275,16 +275,16 @@ class MediterranesnDietView extends StatelessWidget {
                                       ),
                                       child: Column(
                                         mainAxisAlignment:
-                                        MainAxisAlignment.center,
+                                            MainAxisAlignment.center,
                                         crossAxisAlignment:
-                                        CrossAxisAlignment.center,
+                                            CrossAxisAlignment.center,
                                         children: <Widget>[
                                           Text(
                                             '${(1503 * animation!.value).toInt()}',
                                             textAlign: TextAlign.center,
                                             style: TextStyle(
                                               fontFamily:
-                                              FitnessAppTheme.fontName,
+                                                  FitnessAppTheme.fontName,
                                               fontWeight: FontWeight.normal,
                                               fontSize: 24,
                                               letterSpacing: 0.0,
@@ -297,7 +297,7 @@ class MediterranesnDietView extends StatelessWidget {
                                             textAlign: TextAlign.center,
                                             style: TextStyle(
                                               fontFamily:
-                                              FitnessAppTheme.fontName,
+                                                  FitnessAppTheme.fontName,
                                               fontWeight: FontWeight.bold,
                                               fontSize: 12,
                                               letterSpacing: 0.0,
@@ -373,7 +373,7 @@ class MediterranesnDietView extends StatelessWidget {
                                     width: 70,
                                     decoration: BoxDecoration(
                                       color:
-                                      HexColor('#87A0E5').withOpacity(0.2),
+                                          HexColor('#87A0E5').withOpacity(0.2),
                                       borderRadius: BorderRadius.all(
                                           Radius.circular(4.0)),
                                     ),
@@ -406,7 +406,7 @@ class MediterranesnDietView extends StatelessWidget {
                                       fontWeight: FontWeight.w600,
                                       fontSize: 12,
                                       color:
-                                      FitnessAppTheme.grey.withOpacity(0.5),
+                                          FitnessAppTheme.grey.withOpacity(0.5),
                                     ),
                                   ),
                                 ),
@@ -452,7 +452,7 @@ class MediterranesnDietView extends StatelessWidget {
                                               height: 4,
                                               decoration: BoxDecoration(
                                                 gradient:
-                                                LinearGradient(colors: [
+                                                    LinearGradient(colors: [
                                                   HexColor('#F56E98')
                                                       .withOpacity(0.1),
                                                   HexColor('#F56E98'),
@@ -523,7 +523,7 @@ class MediterranesnDietView extends StatelessWidget {
                                               height: 4,
                                               decoration: BoxDecoration(
                                                 gradient:
-                                                LinearGradient(colors: [
+                                                    LinearGradient(colors: [
                                                   HexColor('#F1B440')
                                                       .withOpacity(0.1),
                                                   HexColor('#F1B440'),
